Fix review issues and make clang format
