+++ modules/platforms/cpp/ignite/client/detail/client_error_flags.h
@@ -17,7 +17,6 @@
 
 #pragma once
 
-
 namespace ignite::detail {
 
 /**
