+++ modules/platforms/cpp/ignite/common/CMakeLists.txt
@@ -39,6 +39,7 @@ set(PUBLIC_HEADERS
     ignite_timestamp.h
     ignite_type.h
     primitive.h
+    server_version.h
     uuid.h
 )
