+++ modules/platforms/cpp/ignite/odbc/config/connection_info.cpp
@@ -20,7 +20,7 @@
 #include "ignite/odbc/system/odbc_constants.h"
 #include "ignite/odbc/utility.h"
 
-#include "ignite/common/cluster_version.h"
+#include "ignite/common/server_version.h"
 
 #include <iomanip>
 
@@ -604,7 +604,7 @@ void connection_info::rebuild() {
     // the form ##.##.####, where the first two digits are the major version, the next two digits are the minor version,
     // and the last four digits are the release version. The driver must render the DBMS product version in this form
     // but can also append the DBMS product-specific version. For example, "04.01.0000 Rdb 4.1".
-    m_str_params[SQL_DBMS_VER] = cluster_version().to_string(); // default: 00.00.0000. Will be updated after handshake.
+    m_str_params[SQL_DBMS_VER] = server_version().to_string(); // default: 00.00.0000. Will be updated after handshake.
 #endif // SQL_DBMS_VER
 
 #ifdef SQL_COLUMN_ALIAS
