+++ modules/platforms/cpp/ignite/odbc/sql_connection.cpp
@@ -659,9 +659,9 @@ sql_result sql_connection::make_request_handshake() {
             return sql_result::AI_ERROR;
         }
 
-        auto cluster_ver_str = response.context.get_cluster_version().to_string();
-        LOG_MSG("Cluster version: " << cluster_ver_str);
-        m_info.set_info(SQL_DBMS_VER, cluster_ver_str);
+        auto server_ver_str = response.context.get_server_version().to_string();
+        LOG_MSG("Server version: " << server_ver_str);
+        m_info.set_info(SQL_DBMS_VER, server_ver_str);
 
         auto cluster_name = response.context.get_cluster_name();
         LOG_MSG("Cluster name: " << cluster_name);
